Added Save historical data to file. Needs testing
