+++ DVF-API/Data/Interfaces/IHistoricWeatherDataRepository.cs
@@ -1,10 +1,11 @@
 ﻿using DVF_API.Data.Models;
+using DVF_API.SharedLib.Dtos;
 
 namespace DVF_API.Data.Interfaces
 {
     public interface IHistoricWeatherDataRepository
     {
-        Task SaveDataToFileAsync(WeatherData data, string latitude, string longitude);
+        Task SaveDataToFileAsync(HistoricWeatherDataDto data, string latitude, string longitude, string baseFolder);
         Task SaveDataToDatabaseAsync(WeatherData data);
     }
 }
